refactor: use `which` as alias
