+++ src/commands/what/branch.ts
@@ -4,7 +4,7 @@ import { logger } from '../../context';
 
 export const execute = async () => {
     const branch = await $`git rev-parse --abbrev-ref HEAD`.text();
-    logger.info(string('command.which.branch.action', branch.trim()));
+    logger.info(string('command.what.branch.action', branch.trim()));
 };
 
-export const description = string('command.which.branch.description');
+export const description = string('command.what.branch.description');
