+++ src/strings.ts
@@ -150,7 +150,7 @@ const STRINGS = {
             action: "Pushed them off the stage. Sure hope files can't get hurt.",
             description: "Did you fuckin' stage .env? Remove that!",
         },
-        which: {
+        what: {
             description: chalkTemplate`Are you {italic really} dumb enough to not know info about the project you're literally working on?`,
             branch: {
                 description: chalkTemplate`Really? Did you {bold seriously} forget what branch you're working on? {italic Dumbass}.`,
